Reorganize the way how and when each tab gets displayed
Create a solution that allows to display the next tab
only when all the input data is provided in current
tab. This allows to implement the flow of using the
app - separate calculations and for each of them
create a tab that displays the inteface to provide
the necessary data and select items. Then while
switching tabs, preform calculations and update
the state of the tabs that depend from them.