Fix component validation before updating component data
Introduce a workaround in PreliminaryTabData that, prevents
updating the component data before validating the
inputs to validate.

Without the workaround, if all the data is provided
and user provides a value to the validated input
that is outside the boundaries, since the tab state
checking is perfromed earlier, than input validation,
invalid data will be passed to component data and then
to ShaftDesigner, which might result in error.